refactored scene, background, defaultFog, and defaultLight
