--- lingo3d/src/engine/scene/index.ts
@@ -1,9 +1,0 @@
-import defaultLight from "./defaultLight"
-import defaultFog from "./defaultFog"
-import background from "./background"
-
-preventTreeShake([defaultLight, defaultFog, background])
-
-import scene from "./scene"
-import { preventTreeShake } from "@lincode/utils"
-export default scene
